Updated files with plane assignments
