+++ runfile.py
@@ -1,0 +1,6 @@
+from plane_assign import *
+
+f1 = Flights()
+f1.flightRoutes(short=None, long=None)
+
+
